refactor(agents): one echo block and one config-notes builder, named for their audience
Both config doors were assembling the same stored-vs-effective block field by field and
appending the tier caveat by hand. Two places answering the same question by memory is how
field names drift apart — the shape of the bug this module exists to prevent.

`toolSurfaceEcho` owns the block; `formatConfigSurfaceNotes` owns "every divergence plus the
caveat only a config audience needs", so choosing it over the spawn-side
`formatAgentToolSurfaceNotes` is a deliberate act rather than remembering to append one more
string. `describeSandboxTierCaveat` stops being exported — it has one caller now.
